+++ src/range.ts
@@ -5,7 +5,6 @@
  * Dalliance Genome Explorer, (c) Thomas Down 2006-2010.
  */
 export default class Range {
-
   public ranges: any
   public constructor(arg1: any, arg2?: any) {
     this.ranges = arguments.length === 2 ? [{ min: arg1, max: arg2 }] : 0 in arg1 ? Object.assign({}, arg1) : [arg1]
@@ -44,8 +43,8 @@ export default class Range {
   public union(s1: Range): Range {
     const s0 = this
     const ranges = s0
-      .ranges
-      .concat(s1.ranges)
+      .getRanges()
+      .concat(s1.getRanges())
       .sort(this.rangeOrder)
     const oranges = []
     let current = ranges[0]